Adds exercise 6.
Took me some time to figure out the communication between the caller and its
callback function.

Inside the module, the function definition has as the third parameter another
function, that will GENERATE data and return to its caller (or return an error
if found).

The caller (the main program) defines another function on its side, that
function should use the data returned from the callback defined inside the
imported module.